Flask server is ready
