--- apd/flask/main_flask.py
@@ -1,31 +1,0 @@
-import os
-from flask import Flask, request, render_template
-from werkzeug.utils import secure_filename
-
-dir_path = os.path.dirname(os.path.realpath(__file__))
-
-app = Flask(__name__)
-app.config['UPLOAD_FOLDER'] = os.path.join(dir_path, 'static')
-
-
-@app.route('/pdf')
-def pdf_send():
-    return render_template('index.html')
-
-
-@app.route('/upload')
-def upload_file():
-    return render_template('upload.html')
-
-
-@app.route('/uploader', methods=['GET', 'POST'])
-def upload2_file():
-    if request.method == 'POST':
-        f = request.files['file']
-        f.save(os.path.join(
-            app.config['UPLOAD_FOLDER'], secure_filename(f.filename)))
-        return 'file uploaded successfully'
-
-
-if __name__ == '__main__':
-    app.run()
