+++ apd/flask/templates/upload.html
@@ -1,9 +1,10 @@
 <html>
+
    <body>
-      <form action = "http://localhost:5000/uploader" method = "POST" 
-         enctype = "multipart/form-data">
-         <input type = "file" name = "file" />
-         <input type = "submit"/>
-      </form>   
+      <form action="{{ url_for('getpdf') }}" method="POST" enctype="multipart/form-data">
+         <input type="file" name="file" />
+         <input type="submit" />
+      </form>
    </body>
+
 </html>
