+++ apd/flask/templates/uploader.html
@@ -1,0 +1,11 @@
+<html>
+
+  <head>
+    <meta charset="UTF-8">
+  </head>
+
+  <body>
+    <embed src="{{ file_path }}" height="100%" width="100%" type="application/pdf">
+  </body>
+
+</html>
